pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wqsurrogatemodels"
version = "2.0.0"
description = "WQI5-based current-state water quality assessment backend for WaterMirror."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "kageryo", email = "kageryo@coderyo.com" }
]
dependencies = [
  "fastapi",
  "joblib",
  "loguru",
  "numpy",
  "pandas",
  "python-dotenv",
  "python-multipart",
  "pyyaml",
  "scikit-learn==1.3.2",
  "uvicorn",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "httpx>=0.27.0",
  "anyio>=4.0",
]
test = [
  "pytest>=7.4",
  "httpx>=0.27.0",
  "anyio>=4.0",
]
models = [
  "lightgbm",
  "xgboost",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short -ra"
markers = [
    "anyio: mark test as requiring anyio",
]
